Add `modern_password` to company table
This will eventually replace `password`, which contains the crypted
version.
OGo ObjC will be updated to save a SHA512 of the pwd to that field on
every successful login as well as on password changes.
(LSChangePasswordCommand, LSLoginAccountCommand).

This is two fields and not a crypt=>SHA upgrade, because there may be
a lot of codepathes still just checking for the crypt.